Use existing opacity when calculating fades.
Updates the opacity helper to allow it to understand colors that are
already partially transparent.  This in turn allows the damage flash to
be very dim so that multiple instances of damage in a short time frame
will make the screen glow brightly, but single instances of damage won't
become annoying.